% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa2.R
\name{apa2}
\alias{apa2}
\title{APA: Round a numeric value to 2 decimal places, keeping leading zero}
\usage{
apa2(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
a numeric value
}
\description{
APA: Round a numeric value to 2 decimal places, keeping leading zero
}
\examples{
n <- 1/3
apa2(n)

}
