% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa3.R
\name{apa3}
\alias{apa3}
\title{APA: Round a numeric value to 3 decimal places, keeping leading zero}
\usage{
apa3(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
a numeric value
}
\description{
APA: Round a numeric value to 3 decimal places, keeping leading zero
}
\examples{
n <- 1/3
apa3(n)

}
