% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apan.R
\name{apan}
\alias{apan}
\title{APA: Round a numeric value to \code{n} decimal places, keeping leading zero}
\usage{
apan(x, d)
}
\arguments{
\item{x}{a numeric value}

\item{d}{numbr of decimal places}
}
\value{
a numeric value
}
\description{
APA: Round a numeric value to \code{n} decimal places, keeping leading zero
}
\examples{
n <- 1/3
apa
n2(n, d = 2)

}
