% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstd_to_tibble.R
\name{bstd_to_tibble}
\alias{bstd_to_tibble}
\title{APA: calculate standardized beta estimates for a model}
\usage{
bstd_to_tibble(x, d = 2)
}
\arguments{
\item{x}{REQUIRED: a models (lm or glm), bare name}

\item{d}{Optional: Number. Digits after the decimal place}
}
\value{
a data.frame for use with apaSupp's model tables
}
\description{
APA: calculate standardized beta estimates for a model
}
\examples{

fit_lm1 <- lm(mpg ~ wt + hp + I(factor(cyl)), data = mtcars)

bstd_to_tibble(fit_lm1)

}
