% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta2_to_tibble.R
\name{eta2_to_tibble}
\alias{eta2_to_tibble}
\title{APA: calculate eta-squared effect sizes}
\usage{
eta2_to_tibble(x, d = 2)
}
\arguments{
\item{x}{REQUIRED: a models (lm or glm), bare name}

\item{d}{Optional: Number. Digits after the decimal place}
}
\value{
a data.frame for use with apaSupp's model tables
}
\description{
APA: calculate eta-squared effect sizes
}
\examples{

library(tidyverse)

mtcars <- mtcars \%>\% dplyr::mutate(cyl = factor(cyl))

fit_lm1 <- lm(mpg ~ wt + hp + cyl, data = mtcars)

eta2_to_tibble(fit_lm1)

}
