% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_apa_latex.R
\name{gt_apa_latex}
\alias{gt_apa_latex}
\title{APA Formatted Table of a data.frame in Latex (for .pdf)}
\usage{
gt_apa_latex(x, caption, lab, ...)
}
\arguments{
\item{x}{a data frame}

\item{caption}{text for the caption}

\item{lab}{text for the label}

\item{...}{Other options to pass to gt::gt()}
}
\value{
latex code for a table
}
\description{
APA Formatted Table of a data.frame in Latex (for .pdf)
}
\examples{
df <- head(cars)
gt_apa_latex(df, caption = "Testing", lab = "tab1")
}
