% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_apa_screen.R
\name{gt_apa_screen}
\alias{gt_apa_screen}
\title{APA Formatted Table of a data.frame in non-Latex (for screen}
\usage{
gt_apa_screen(x, caption, ...)
}
\arguments{
\item{x}{a data frame}

\item{caption}{text for the caption}

\item{...}{Other options to pass to gt::gt()}
}
\value{
a table
}
\description{
APA Formatted Table of a data.frame in non-Latex (for screen
}
\examples{
df <- head(cars)
gt_apa_screen(df, caption = "Testing")
}
