% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_gee.R
\name{gt_gee}
\alias{gt_gee}
\title{APA: gtsummary for a single GEE model}
\usage{
gt_gee(x, narrow = FALSE, fit = NA, d = 2, show_single_row = NULL)
}
\arguments{
\item{x}{REQUIRED: bare name. a single 'geeglm' object'}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: vector. quoted names of fit statistics to include, can be: "r.squared", "adj.r.squared", "sigma", "statistic","p.value", "df", "logLik", "AIC", "BIC", "deviance", "df.residual", and "nobs"}

\item{d}{Optional: number. digits after the decimal, default = 2}

\item{show_single_row}{(tidy-select) By default categorical variables are printed on multiple rows. If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}
}
\value{
a gtsummary object
}
\description{
APA: gtsummary for a single GEE model
}
\examples{

library(HSAUR)
library(tidyverse)
library(geepack)

data("respiratory", package = "HSAUR")

data_wide <- respiratory \%>\%
 tidyr::spread(key = month,
               value = status,
               sep = "_") \%>\%
 dplyr::rename("BL_status" = "month_0") \%>\%
 dplyr::arrange(subject) \%>\%
 dplyr::select(subject, centre,
               sex, age, treatment,
               BL_status, starts_with("month"))

data_long <- data_wide\%>\%
 tidyr::gather(key = month,
               value = status,
               starts_with("month")) \%>\%
 dplyr::mutate(month = str_sub(month, start = -1) \%>\% as.numeric) \%>\%
 dplyr::mutate(status = case_when(status == "poor" ~ 0,
                                  status == "good" ~ 1)) \%>\%
 dplyr::arrange(subject, month) \%>\%
 dplyr::select(subject, centre, sex, age, treatment, BL_status, month, status)


resp_geeglm_ex <- geepack::geeglm(status ~ centre + treatment + sex + BL_status +
                                   I(age-33) + I((age-33)^2),
                                 data = data_long,
                                 family = binomial(link = "logit"),
                                 id = subject,
                                 waves = month,
                                 corstr = "exchangeable")

gt_gee(resp_geeglm_ex)


resp_geeglm2_ex <- geepack::geeglm(status ~ treatment + sex + BL_status,
                                  data = data_long,
                                  family = binomial(link = "logit"),
                                  id = subject,
                                  waves = month,
                                  corstr = "exchangeable")

gt_gee(resp_geeglm2_ex, narrow = TRUE)


data(BtheB, package = "HSAUR")

btb_wide <- BtheB \%>\%
 dplyr::mutate(id = row_number()) \%>\%           # create a new variable to ID participants
 dplyr::select(id, treatment,                    # specify that ID variable is first
               drug, length,
               bdi.pre, bdi.2m, bdi.4m, bdi.6m, bdi.8m)

btb_long <- btb_wide \%>\%
 tidyr::pivot_longer(cols = c(bdi.2m, bdi.4m, bdi.6m, bdi.8m),  # all existing variables (not quoted)
                     names_to = "month",
                     names_pattern = "bdi.(.)m",
                     values_to = "bdi") \%>\%
 dplyr::mutate(month = as.numeric(month)) \%>\%
 dplyr::filter(complete.cases(id, bdi, treatment, month)) \%>\%
 dplyr::arrange(id, month) \%>\%
 dplyr::select(id, treatment, drug, length, bdi.pre, month, bdi)

btb_geeglm_ex_1 <- geepack::geeglm(bdi ~ bdi.pre*length + drug + treatment + month,
                                  data = btb_long,
                                  id = id,
                                  wave = month,
                                  family = gaussian,
                                  corstr = 'exchangeable')


gt_gee(btb_geeglm_ex_1)
gt_gee(btb_geeglm_ex_1, narrow = TRUE)

}
