% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_glm.R
\name{gt_glm}
\alias{gt_glm}
\title{APA: gtsummary for a single linear model}
\usage{
gt_glm(x, narrow = FALSE, fit = c("r.squared", "adj.r.squared"), d = 2)
}
\arguments{
\item{x}{REQUIRED: bare name. a single 'glm' object'}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: vector. quoted names of fit statistics to include, can be: "r.squared", "adj.r.squared", "sigma", "statistic","p.value", "df", "logLik", "AIC", "BIC", "deviance", "df.residual", and "nobs"}

\item{d}{Optional: number. digits after the decimal, default = 2}
}
\value{
a gtsummary object
}
\description{
APA: gtsummary for a single linear model
}
\examples{
library(datasets)
library(tidyverse)

mtcars <- mtcars \%>\% dplyr::mutate(vs = factor(vs))
m <- lm(mpg ~ vs + disp, data = mtcars)
gt_lm(m)



}
