% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_glmer.R
\name{gt_glmer}
\alias{gt_glmer}
\title{APA: gtsummary for a single generalized linear mioxed effects model}
\usage{
gt_glmer(
  x,
  narrow = FALSE,
  fit = c("AIC", "BIC", "logLik"),
  d = 2,
  show_single_row = NULL
)
}
\arguments{
\item{x}{REQUIRED: bare name. a single 'glmer' object'}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: vector. quoted names of fit statistics to include, can be: "AIC", "BIC", "logLik"}

\item{d}{Optional: number. digits after the decimal, default = 2}

\item{show_single_row}{(tidy-select) By default categorical variables are printed on multiple rows. If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}
}
\value{
a gtsummary object
}
\description{
APA: gtsummary for a single generalized linear mioxed effects model
}
\examples{

library(tidyverse)
library(lme4)
library(lmerTest)
library("HSAUR3")

fit1 <- glmer(outcome~treatment+(1|patientID), data=toenail,family=binomial,nAGQ=20)
fit2 <- glmer(outcome~treatment+visit+(1|patientID), data=toenail,family=binomial,nAGQ=20)
fit3 <- glmer(outcome~treatment*visit+(1|patientID), data=toenail,family=binomial,nAGQ=20)


apaSupp::gt_glmer(fit1)
apaSupp::gt_glmer(fit2)
apaSupp::gt_glmer(fit3)


}
