% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_lm.R
\name{gt_lm}
\alias{gt_lm}
\title{APA: gtsummary for a single linear model}
\usage{
gt_lm(x, d = 2)
}
\arguments{
\item{x}{REQUIRED: a single 'lm' object, bare name}

\item{d}{OPTIONAL: number of digits, default = 2}
}
\value{
a gtsummary object
}
\description{
APA: gtsummary for a single linear model
}
\examples{
m <- lm(dist ~ speed, data = cars)
gt_lm(m)



}
