% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_lmer.R
\name{gt_lmer}
\alias{gt_lmer}
\title{APA: gtsummary for a single linear mixed effects model}
\usage{
gt_lmer(
  x,
  narrow = FALSE,
  fit = c("AIC", "BIC", "logLik"),
  d = 2,
  show_single_row = NULL
)
}
\arguments{
\item{x}{REQUIRED: bare name. a single 'lmer' object'}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: Character vector. Default  = c("AIC", "BIC", "logLik).  May include quoted names of fit statistics listed from broom::glimpse(x)}

\item{d}{Optional: number. digits after the decimal, default = 2}

\item{show_single_row}{(tidy-select) By default categorical variables are printed on multiple rows. If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}
}
\value{
a gtsummary object
}
\description{
APA: gtsummary for a single linear mixed effects model
}
\examples{

library(tidyverse)
library(lme4)
library(lmerTest)

fm1 <- lmerTest::lmer(Reaction ~ Days + (1 | Subject), sleepstudy)
fm2 <- lmerTest::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
fm3 <- lmerTest::lmer(Reaction ~ Days + (1 | Subject), sleepstudy \%>\% dplyr::mutate(Days = factor(Days)))

apaSupp::gt_lmer(fm1)
apaSupp::gt_lmer(fm2, d = 3)
apaSupp::gt_lmer(fm3, narrow = TRUE)

}
