% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_sd.R
\name{lmer_sd}
\alias{lmer_sd}
\title{Estimate the SD for Effect Size for a MLM}
\usage{
lmer_sd(fit, type = "pool")
}
\arguments{
\item{fit}{REQUIRED. Bare name of a model fit via lme4::lmer()}

\item{type}{REQUIRED. Text. Default = "pool", alternative is "resid"}
}
\value{
list of 2 estimates of standard deviation
}
\description{
Estimate the SD for Effect Size for a MLM
}
\examples{

library(tidyverse)
library(lme4)

fit <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy)

lmer_sd(fit)
lmer_sd(fit, type = "resid")

}
