% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_num.R
\name{p_num}
\alias{p_num}
\title{P-value Numbers: 3-decimal places, no leading zero, and stars}
\usage{
p_num(
  value,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***"),
  decimals = 3,
  leading = FALSE
)
}
\arguments{
\item{value}{a numeric value for a statistical significance (p-value)}

\item{breaks}{statistical significance break points}

\item{symbols}{symbols to assign to each break point}

\item{decimals}{number of digits after the decimal point}

\item{leading}{keep the leading zero in front of the decimal}
}
\value{
character of the form ".231", ".022**", or "< .001***"
}
\description{
P-value Numbers: 3-decimal places, no leading zero, and stars
}
\examples{
x <- c(.36521, .02456, .0000000056)
p_num(x)
}
