% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_star.R
\name{p_star}
\alias{p_star}
\title{P-value Stars: only stars with no numbers}
\usage{
p_star(value, breaks = c(0.05, 0.01, 0.001), symbols = c("*", "**", "***"))
}
\arguments{
\item{value}{a numeric value for a statistical significance (p-value)}

\item{breaks}{statistical significance break points}

\item{symbols}{symbols to assign to each break point}
}
\value{
character of the form ".231", ".022**", or "< .001***"
}
\description{
P-value Stars: only stars with no numbers
}
\examples{

x <- c(.36521, .02456, .0000000056)

p_star(x)

}
