% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plrt_to_tibble.R
\name{plrt_to_tibble}
\alias{plrt_to_tibble}
\title{APA: calculate significance for single term deletion}
\usage{
plrt_to_tibble(x, test = "Chisq", d = 2, k = 2)
}
\arguments{
\item{x}{REQUIRED: a models (lm or glm), bare name}

\item{test}{Optional: Character.  Default = "Chisq", which is an exact test (\code{lm} models with known scale) or a likelihood-ratio test or a test of the reduction in scaled deviance \code{glm} models, see help for \code{drop1()}}

\item{d}{Optional: Number. Digits after the decimal place}
}
\value{
a data.frame for use with apaSupp's model tables
}
\description{
APA: calculate significance for single term deletion
}
\examples{

fit_lm1 <- lm(mpg ~ wt + hp + I(factor(cyl)), data = mtcars)
fit_glm1 <-

plrt_to_tibble(fit_lm1)

}
