% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spicy_histo.R
\name{spicy_histo}
\alias{spicy_histo}
\title{Descriptive Summary of a Continuous Variable, Histogram and Table}
\usage{
spicy_histo(
  df,
  var,
  lab = NULL,
  binwidth = NULL,
  bins = 10,
  digits = 2,
  fontname = "serif"
)
}
\arguments{
\item{df}{REQUIRED: data frame}

\item{var}{REQUIRED: Bare variable name.}

\item{lab}{Optional: Text. Label for x-axis of plot}

\item{binwidth}{Optional: Number. Width of bins}

\item{bins}{Optional: Number. Quantity of bins (default = 10)}

\item{digits}{Optional: Number. Digits after the decimal place}

\item{fontname}{Optional: Text.  Font used in table}
}
\value{
a \code{ggplot} object
}
\description{
Descriptive Summary of a Continuous Variable, Histogram and Table
}
\examples{
# Simple
spicy_histo(mtcars, mpg)

# More Complex
spicy_histo(df = mtcars,
            var = mpg,
            lab = "Fuel Economy, mpg",
            binwidth = 2)

}
