% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spicy_histos.R
\name{spicy_histos}
\alias{spicy_histos}
\title{Descriptive Summary of a Continuous Variable, Histograms and Table}
\usage{
spicy_histos(
  df,
  var,
  split = NULL,
  lab = NULL,
  binwidth = NULL,
  bins = 10,
  digits = 2
)
}
\arguments{
\item{df}{REQUIRED: data frame}

\item{var}{REQUIRED: Bare variable name. Variable to compare}

\item{split}{REQUIRED: Bare variable name. Row by a factor}

\item{lab}{Optional: Text. Label for x-axis of plot}

\item{binwidth}{Optional: Number. Width of bins}

\item{bins}{Optional: Number. Quantity of bins (default = 10)}

\item{digits}{Optional: Number. Digits after the decimal place for numeric values}
}
\value{
a \code{ggplot} object
}
\description{
Descriptive Summary of a Continuous Variable, Histograms and Table
}
\examples{
# Simple
library(datasets)

 spicy_histos(df = mtcars,
              var = mpg,
              split = am)

# Complex
spicy_histos(df = mtcars,
             var = mpg,
             lab = "Fuel Economy, mpg",
             split = vs,
             binwidth = 2)

}
