% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spicy_violins.R
\name{spicy_violins}
\alias{spicy_violins}
\title{Descriptive Summary of a Continuous Variable, Violins and Table}
\usage{
spicy_violins(
  df,
  var,
  lab = NULL,
  split = NULL,
  digits = 2,
  fontname = "serif",
  heights = c(5, 1.5)
)
}
\arguments{
\item{df}{REQUIRED: data frame}

\item{var}{REQUIRED: Bare variable name.}

\item{lab}{Optional: Text. Label for x-axis of plot}

\item{split}{REQUIRED: Bare variable name. Row by a factor}

\item{digits}{Optional: Number. Digits after the decimal place}

\item{fontname}{Optional: Text.  Font used in table}

\item{heights}{Optional: Vector of 2 numbers. figure vs. table}
}
\value{
a \code{ggplot} object
}
\description{
Descriptive Summary of a Continuous Variable, Violins and Table
}
\examples{

spicy_violins(df = mtcars,
              var = mpg,
              lab = "Fuel Economy, mpg",
              split = vs)

spicy_violins(df = chickwts,
              var = weight,
              lab = "Chick Weight at Six Weeks (g)",
              split = feed,
              heights = c(2,1))

}
