% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\title{Descriptive Table, total or split by a factor}
\usage{
tab1(
  df,
  split = NULL,
  caption = "Summary of Variables",
  docx = NA,
  no_notes = FALSE,
  general_note = NA,
  p_note = "apa123",
  total = TRUE,
  var_lab = " ",
  test = TRUE,
  total_lab = "Total\\nN = {N}",
  total_last = FALSE,
  na.rm = FALSE,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***"),
  d = 2,
  statistic = list(all_continuous() ~ "{mean} ({sd})", all_categorical() ~
    "{n}    ({p}\%)"),
  value = NULL,
  type = NULL,
  missing = "always",
  missing_text = "Unknown",
  missing_stat = "{N_miss}",
  sort = all_categorical(FALSE) ~ "alphanumeric",
  percent = c("column", "row", "cell"),
  include = everything()
)
}
\arguments{
\item{df}{(\code{data.frame})\cr A data frame.}

\item{split}{Optional: Quoted variable name}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: Text.  Word file to store the table.  Must end with ".docx"}

\item{no_notes}{Optional: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa123"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{total}{Optional: Logical. Default = TRUE to include a total column in a split table}

\item{var_lab}{Optional: Text. Label to print above the variable names in the header}

\item{test}{Optional: Logical. Default = TRUE to run comparisons in a split table}

\item{total_lab}{Optional: Text. Label above the total column in a split table}

\item{total_last}{Optional: Logical. Default = FALSE to include the total column last vs first}

\item{na.rm}{Optional: Logical. Default = FALSE, do not remove instances with a missing value from the entire table}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}

\item{d}{Optional: Number. Digits after the decimal place}

\item{statistic}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Used to specify the summary statistics for each variable.
The default is
\code{list(all_continuous()  ~ "{mean} ({sd})", all_categorical() ~ "{n}    ({p}\%)")}.}

\item{value}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies the level of a variable to display on a single row.
The gtsummary type selectors, e.g. \code{all_dichotomous()}, cannot be used
with this argument. Default is \code{NULL}.}

\item{type}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies the summary type. Accepted value are
\code{c("continuous", "continuous2", "categorical", "dichotomous")}.
If not specified, default type is assigned via
\code{assign_summary_type()}.}

\item{missing, missing_text, missing_stat}{Arguments dictating how and if missing values are presented:
\itemize{
\item \code{missing}: must be one of \code{c("ifany", "no", "always")}.
\item \code{missing_text}: string indicating text shown on missing row. Default is \code{"Unknown"}.
\item \code{missing_stat}: statistic to show on missing row. Default is \code{"{N_miss}"}.
Possible values are \code{N_miss}, \code{N_obs}, \code{N_nonmiss}, \code{p_miss}, \code{p_nonmiss}.
}}

\item{sort}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies sorting to perform for categorical variables.
Values must be one of \code{c("alphanumeric", "frequency")}.
Default is \code{all_categorical(FALSE) ~ "alphanumeric"}.}

\item{percent}{(\code{string})\cr
Indicates the type of percentage to return.
Must be one of \code{c("column", "row", "cell")}. Default is \code{"column"}.

In rarer cases, you may need to define/override the typical denominators.
In these cases, pass an integer or a data frame. Refer to the
\code{\link[cards:ard_categorical]{?cards::ard_categorical(denominator)}} help file for details.}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to include in the summary table. Default is \code{everything()}.}
}
\value{
a \code{flextable} table with caption
}
\description{
Descriptive Table, total or split by a factor
}
\examples{
library(tidyverse)
library(gtsummary)

data(mtcars, package = "datasets")

mtcars <- mtcars \%>\%
  dplyr::mutate(cyl = factor(cyl)) \%>\%
  dplyr::mutate(vs = factor(vs,
                            levels = c(0, 1),
                            labels = c("V-shaped", "Straight"))) \%>\%
  dplyr::mutate(am = factor(am,
                            levels = c(0, 1),
                            labels = c("Automatic",
                                       "Manual")))


##########################

tab1(mtcars)


mtcars \%>\%
  dplyr::select(cyl,
                "Fuel Economy, mpg" = mpg,
                "Dispacement, cu-in" = disp,
                "Gross Horsepower" = hp,
                "Real Axel Ratio" = drat,
                "Weight, 1000 lbs" = wt,
                "Engine Type" = vs) \%>\%
  tab1(split        = "cyl",
       total_last   = FALSE,
       caption      = "Descriptive Summary of Car Road Tests",
       general_note = "Data extracted from the 1974 Motor Trend US magazine.",
       p_note       = "apa3")
}
