% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_apa.R
\name{tab_apa}
\alias{tab_apa}
\title{APA Formatted Table of a data.frame}
\usage{
tab_apa(
  df,
  caption,
  general_note = NULL,
  p_note = "* p < .05. ** p < .01. *** p < .001.",
  valign = "center",
  digits = 2,
  fontname = "serif",
  space = 0.5
)
}
\arguments{
\item{df}{REQUIRED: data frame}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. Significance note for APA table}

\item{valign}{Optional: Text. vertical alignment ("center")}

\item{digits}{Optional: Number. Digits after the decimal place}

\item{fontname}{Optional: Text.  Font used in table}

\item{space}{Optional: Number. Line spacing in the body of the table}
}
\value{
table
}
\description{
APA Formatted Table of a data.frame
}
\examples{
library(tidyverse)
library(psych)

tab_apa(head(cars), caption = "Cars Dataset")

cars \%>\%
head() \%>\%
tab_apa(caption = "Speed and Stopping Distances of Cars in the 1920's",
        general_note = "Example cases from the dataset.",
        p_note = NULL)

cars \%>\%
  psych::headTail() \%>\%
  tibble::rownames_to_column(var = "ID") \%>\%
  dplyr::rename("Speed, mph" = speed,
                "Stopping Distance, ft" = dist) \%>\%
  tab_apa(caption = "Speed and Stopping Distances of Cars in the 1920's",
          general_note = "Head and tail of a dataset from:
             McNeil, D. R. (1977) Interactive Data Analysis. Wiley.",
          p_note = NULL)
}
