% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_desc.R
\name{tab_desc}
\alias{tab_desc}
\title{Basic Descriptive Summary: n, M (SD), \link{Min, Max}, Median with flextable}
\usage{
tab_desc(df, caption, lab = "Scale")
}
\arguments{
\item{df}{a data frame with relevant variables selected}

\item{caption}{the \code{flextable} caption for the table}

\item{lab}{the text to place above the variable names}
}
\value{
a \code{flextable} table with caption
}
\description{
Basic Descriptive Summary: n, M (SD), \link{Min, Max}, Median with flextable
}
\examples{
library(tidyverse)

tab_desc(df = cars, caption = "Demographics Summary of Cars", lab = "Variable")

cars \%>\%
dplyr::rename("Speed, mph" = speed,
"Stopping Distance, ft" = dist) \%>\%
tab_desc(caption = "Demographics Summary of Cars", lab = "Variable")

}
