% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_freq.R
\name{tab_freq}
\alias{tab_freq}
\title{Frequency Table, total or split by a factor}
\usage{
tab_freq(
  df,
  split = NULL,
  caption,
  general_note = NULL,
  max_width_in = 6,
  digits = 2
)
}
\arguments{
\item{df}{REQUIRED: Data frame}

\item{split}{Optional: Quoted variable name}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{max_width_in}{Optional: Number.  Inches wide table can be}

\item{digits}{Optional: Number. Digits after the decimal place}
}
\value{
a \code{flextable} table with caption
}
\description{
Frequency Table, total or split by a factor
}
\examples{
library(tidyverse)

mtcars
}
