% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_gee.R
\name{tab_gee}
\alias{tab_gee}
\title{APA: flextable for a GEE models (specifically logisitic right now)}
\usage{
tab_gee(
  x,
  var_labels = NULL,
  caption = "Parameter Estimates for Generalized Linear Regression",
  docx = NA,
  tab_width = 0.9,
  general_note = NA,
  p_note = "apa123",
  no_notes = FALSE,
  d = 2,
  show_single_row = NULL,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***")
)
}
\arguments{
\item{x}{REQUIRED: a gee models, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{tab_width}{Optional: numeric value (default is .9) \% of available width}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa123"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{show_single_row}{a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}
}
\value{
a flextable object
}
\description{
APA: flextable for a GEE models (specifically logisitic right now)
}
\examples{


library(HSAUR)
library(tidyverse)
library(gee)
library(geepack)

data("respiratory", package = "HSAUR")


data_long <- data_wide\%>\%
 tidyr::gather(key = month,
               value = status,
               starts_with("month")) \%>\%
 dplyr::mutate(month = str_sub(month, start = -1) \%>\% as.numeric) \%>\%
 dplyr::mutate(status = case_when(status == "poor" ~ 0,
                                  status == "good" ~ 1)) \%>\%
 dplyr::arrange(subject, month) \%>\%
 dplyr::select(subject, centre, sex, age, treatment, BL_status, month, status)


resp_geeglm_ex <- geepack::geeglm(status ~ centre + treatment + sex + BL_status +
                                   I(age-33) + I((age-33)^2),
                                 data = data_long,
                                 family = binomial(link = "logit"),
                                 id = subject,
                                 waves = month,
                                 corstr = "exchangeable")

tab_gee(resp_geeglm_ex)


resp_geeglm2_ex <- geepack::geeglm(status ~ treatment + sex + BL_status,
                                  data = data_long,
                                  family = binomial(link = "logit"),
                                  id = subject,
                                  waves = month,
                                  corstr = "exchangeable")

tab_gee(resp_geeglm2_ex)



}
