% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_glm.R
\name{tab_glm}
\alias{tab_glm}
\title{APA: flextable for a GLM models (specifically logisitc right now)}
\usage{
tab_glm(
  x,
  var_labels = NULL,
  caption = "Generalized Regression Model",
  p_note = "apa",
  general_note = NA,
  fit = c("null.deviance", "deviance"),
  d = 2,
  vif = TRUE
)
}
\arguments{
\item{x}{REQUIRED: a glm models, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If p_note = "apa" then the standard "* p < .05. ** p < .01. *** p < .001." will be used}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{fit}{Optional: Text. fit statistics:  "nobs", null.deviance", "df.null", "deviance", "df.residual", "logLik", "AIC", "BIC"}

\item{d}{Optional: Number. Digits after the decimal place}

\item{type}{Optional: default = "logistic", more to come soon...}
}
\value{
a flextable object
}
\description{
APA: flextable for a GLM models (specifically logisitc right now)
}
\examples{


mtcars <- mtcars \%>\% dplyr::mutate(cyl = factor(cyl))

fit_glm2 <- glm(vs ~ wt + mpg + cyl, data = mtcars, family = "binomial")

apaSupp::tab_glm(fit_glm2)

}
