% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_glm_exp.R
\name{tab_glm_exp}
\alias{tab_glm_exp}
\title{APA: flextable for a GLM models (specifically logisitc right now)}
\usage{
tab_glm_exp(
  x,
  var_labels = NULL,
  type = "logistic",
  caption = "Generalized Regression Model",
  p_note = "apa",
  general_note = NA,
  d = 2
)
}
\arguments{
\item{x}{REQUIRED: a glm models, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{type}{Optional: default = "logistic", more to come soon...}

\item{caption}{Optional: Text. Caption for the table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If p_note = "apa" then the standard "* p < .05. ** p < .01. *** p < .001." will be used}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{d}{Optional: Number. Digits after the decimal place}
}
\value{
a flextable object
}
\description{
APA: flextable for a GLM models (specifically logisitc right now)
}
\examples{
library(HSAUR3)
library(tidyverse)

data("womensrole", package = "HSAUR3")

fit_glm_1 <- glm(cbind(agree, disagree) ~ gender + education,
                  data = womensrole,
                  family = binomial)

apaSupp::tab_glm_exp(fit_glm_1)

}
