% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_glmer.R
\name{tab_glmer}
\alias{tab_glmer}
\title{APA: flextable for ONE generarlized linear mixed effects models}
\usage{
tab_glmer(
  x,
  var_labels = NULL,
  caption = "Parameter Estimates for Generalized Mixed Effects Regression",
  docx = NA,
  general_note = NA,
  p_note = "apa123",
  no_notes = FALSE,
  d = 2,
  ci = FALSE,
  show_single_row = NULL,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***")
)
}
\arguments{
\item{x}{REQUIRED: 1 glmer model, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa123"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{ci}{Optional: logical. (default = FALSE) Include a confidence interval for the estimated beta}

\item{show_single_row}{Optional: If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}
}
\value{
a flextable object
}
\description{
APA: flextable for ONE generarlized linear mixed effects models
}
\examples{

library(tidyverse)
library(lme4)
library(lmerTest)
library("HSAUR3")

fit1 <- glmer(outcome~treatment*visit+(1|patientID), data=toenail,family=binomial,nAGQ=20)



apaSupp::tab_glmer(fit1)


apaSupp::tab_glmer(fit1, show_single_row = c("treatment","treatment:visit"), docx = "test.docx")

}
