% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_glmers.R
\name{tab_glmers}
\alias{tab_glmers}
\title{APA: flextable for 2-3 generalized linear mixed effects models}
\usage{
tab_glmers(
  x,
  var_labels = NULL,
  caption = "Compare Generalized Mixed-Effects Models",
  docx = NA,
  general_note = NA,
  p_note = "apa123",
  no_notes = FALSE,
  d = 2,
  narrow = FALSE,
  fit = c("AIC", "BIC", "logLik"),
  show_single_row = NULL,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***")
)
}
\arguments{
\item{x}{REQUIRED: List. at least 2 glmer models, bare names}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If p_note = "apa" then the standard "* p < .05. ** p < .01. *** p < .001." will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: vector. quoted names of fit metrics from \code{performance::performance()} for glmer (max of 4)}

\item{show_single_row}{(tidy-select) By default categorical variables are printed on multiple rows. If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}
}
\value{
a flextable object
}
\description{
APA: flextable for 2-3 generalized linear mixed effects models
}
\examples{

library(tidyverse)
library(lme4)
library(lmerTest)
library("HSAUR3")

fit1 <- glmer(outcome~treatment+(1|patientID), data=toenail,family=binomial,nAGQ=20)
fit2 <- glmer(outcome~treatment+visit+(1|patientID), data=toenail,family=binomial,nAGQ=20)
fit3 <- glmer(outcome~treatment*visit+(1|patientID), data=toenail,family=binomial,nAGQ=20)

apaSupp::tab_glmers(list(fit1, fit2))
apaSupp::tab_glmers(list(fit1, fit2, fit3), narrow = TRUE)



}
