% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_glms.R
\name{tab_glms}
\alias{tab_glms}
\title{APA: flextable for 2-3 generalized linear models}
\usage{
tab_glms(
  x,
  var_labels = NULL,
  caption = "Compare Generalized Regression Models",
  docx = NA,
  tab_width = 0.9,
  general_note = NA,
  p_note = "apa123",
  no_notes = FALSE,
  d = 2,
  narrow = FALSE,
  fit = c("AIC", "BIC"),
  pr2 = "both",
  show_single_row = NULL,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***")
)
}
\arguments{
\item{x}{REQUIRED: List. at least 2 glm models, bare names}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{tab_width}{Optional: numberic value (default is .9) \% of available width}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If p_note = "apa" then the standard "* p < .05. ** p < .01. *** p < .001." will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: vector. quoted names of fit metrics from \code{performance::performance()} for glm (max of 4)}

\item{pr2}{Optional: character.  (default = "both") Include pseudo R-squared: "tjur", "mcfadden", "both", or "none"}

\item{show_single_row}{(tidy-select) By default categorical variables are printed on multiple rows. If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}
}
\value{
a flextable object
}
\description{
APA: flextable for 2-3 generalized linear models
}
\examples{

library(tidyverse)

mtcars <- mtcars \%>\% dplyr::mutate(vs = factor(vs))

fit_glm1 <- glm(vs ~ wt, data = mtcars, family = "binomial")
fit_glm2 <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")

apaSupp::tab_glms(list(fit_glm1, fit_glm2))
apaSupp::tab_glms(list("M1" = fit_glm1, "M2" = fit_glm2), pr2 = "both")
apaSupp::tab_glms(list("M1" = fit_glm1, "M2" = fit_glm2), narrow = TRUE)
apaSupp::tab_glms(list("M1" = fit_glm1, "M2" = fit_glm2), narrow = TRUE, fit = "RMSE")
apaSupp::tab_glms(list("M1" = fit_glm1, "M2" = fit_glm2), narrow = TRUE, fit = NA)
apaSupp::tab_glms(list("M1" = fit_glm1, "M2" = fit_glm2), narrow = TRUE, fit = NA, pr2 = "tjur")
apaSupp::tab_glms(list("M1" = fit_glm1, "M2" = fit_glm2), narrow = TRUE, fit = "AIC", pr2 = "tjur")


fit_glm3 <- glm(vs ~ wt + mpg, data = mtcars[1:30,], family = "binomial")

apaSupp::tab_glms(list(fit_glm1, fit_glm2, fit_glm3), narrow = TRUE)
apaSupp::tab_glms(list(fit_glm1, fit_glm2, fit_glm3), narrow = TRUE, fit = "AIC", pr2 = "tjur")



}
