% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lm_fits.R
\name{tab_lm_fits}
\alias{tab_lm_fits}
\title{APA: flextable for Comparing the Performance of Linear models}
\usage{
tab_lm_fits(
  x,
  caption = "Comparison of Linear Model Performane Metrics",
  general_note = NA,
  sort = "AIC",
  d = 2
)
}
\arguments{
\item{x}{REQUIRED: List. at least 2 lm models, bare names, If named list, then names appear in the table}

\item{caption}{Optional: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{sort}{Optional: metrics to sort by, default = "AIC", but may use: "AIC", "BIC", "R2", "R2_adjusted", "RMSE"}

\item{d}{Optional: Number. Digits after the decimal place}
}
\value{
a flextable object
}
\description{
Create a flextable for Comparing the Performance of Linear models via Several Metrics
}
\examples{

m1 <- lm(dist ~ 1, cars)
m2 <- lm(dist ~ speed, cars)

apaSupp::tab_lm_fits(list(m1, m2))
apaSupp::tab_lm_fits(list("null" = m1, "main" = m2))

}
