% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lmer.R
\name{tab_lmer}
\alias{tab_lmer}
\title{APA: flextable for ONElinear models}
\usage{
tab_lmer(
  x,
  var_labels = NULL,
  caption = "Parameter Estimates for Mixed Effects Regression",
  docx = NA,
  tab_width = 0.9,
  general_note = NA,
  p_note = "apa123",
  no_notes = FALSE,
  d = 2,
  ci = FALSE,
  show_single_row = NULL,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***")
)
}
\arguments{
\item{x}{REQUIRED: 1 lm model, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{tab_width}{Optional: numberic value (default is .9) \% of available width}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa123"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{ci}{Optional: logical. (default = FALSE) Include a confidence interval for the estimated beta}

\item{show_single_row}{Optional: If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}
}
\value{
a flextable object
}
\description{
APA: flextable for ONElinear models
}
\examples{

library(tidyverse)
library(lme4)
library(lmerTest)

fm1 <- lmerTest::lmer(Reaction ~ Days + (1 | Subject), sleepstudy)
fm2 <- lmerTest::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
fm3 <- lmerTest::lmer(Reaction ~ Days + (1 | Subject), sleepstudy \%>\% dplyr::mutate(Days = factor(Days)))

apaSupp::tab_lmer(fm1)


apaSupp::tab_lmer(fm2, ci = TRUE, docx = "test.docx")

}
