% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lmer_fits.R
\name{tab_lmer_fits}
\alias{tab_lmer_fits}
\title{APA: flextable for Comparing the Performance of Linear Mixed Effects models}
\usage{
tab_lmer_fits(
  x,
  caption = "Comparison of MLM Performane Metrics",
  docx = NA,
  general_note = NA,
  d = 2
)
}
\arguments{
\item{x}{REQUIRED: List. at least 2 lm models, bare names, If named list, then names appear in the table}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{d}{Optional: Number. Digits after the decimal place}
}
\value{
a flextable object
}
\description{
Create a flextable for Comparing the Performance of MLMs via Several Metrics
}
\examples{

library(tidyverse)
library(lme4)
library(lmerTest)

fm1 <- lmerTest::lmer(Reaction ~ Days + (1 | Subject), sleepstudy)
fm2 <- lmerTest::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
fm3 <- lmerTest::lmer(Reaction ~ Days + (1 | Subject), sleepstudy \%>\% dplyr::mutate(Days = factor(Days)))
fm4 <- lmerTest::lmer(Reaction ~ Days + I(Days^2) + (1|Subject), sleepstudy)


apaSupp::tab_lmer_fits(list(fm1, fm2, fm3, fm4))



}
