% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lmers.R
\name{tab_lmers}
\alias{tab_lmers}
\title{APA: flextable for 2-3 linear mixed effects models}
\usage{
tab_lmers(
  x,
  var_labels = NULL,
  caption = "Compare MLM Models",
  docx = NA,
  tab_width = 0.9,
  general_note = NA,
  p_note = "apa123",
  no_notes = FALSE,
  d = 2,
  narrow = FALSE,
  fit = c("AIC", "BIC", "logLik"),
  show_single_row = NULL,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***")
)
}
\arguments{
\item{x}{REQUIRED: List. at least 2 lmer models, bare names}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{tab_width}{Optional: numberic value (default is .9) \% of available width}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If p_note = "apa" then the standard "* p < .05. ** p < .01. *** p < .001." will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: vector. quoted names of fit statistics to include, can be: "AIC", "BIC", "logLik", "statistic","p.value", "df", "logLik", "AIC", "BIC", "deviance", "df.residual", and "nobs"}

\item{show_single_row}{If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}
}
\value{
a flextable object
}
\description{
APA: flextable for 2-3 linear mixed effects models
}
\examples{

library(tidyverse)
library(lme4)
library(lmerTest)

fm1 <- lmerTest::lmer(Reaction ~ Days + (1 | Subject), sleepstudy)
fm2 <- lmerTest::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
fm3 <- lmerTest::lmer(Reaction ~ Days + (1 | Subject), sleepstudy \%>\% dplyr::mutate(Days = factor(Days)))
fm4 <- lmerTest::lmer(Reaction ~ Days + I(Days^2) + (1|Subject), sleepstudy)

apaSupp::tab_lmers(list(fm1, fm2))
apaSupp::tab_lmers(list(fm1, fm2, fm4), var_labels = c("Subject.var__(Intercept)" = "rand int"))

}
