% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_apa.R
\name{table1_apa}
\alias{table1_apa}
\title{Convert a \code{furniture::table1()} to a \code{flextable::flextable()} with APA formatting}
\usage{
table1_apa(
  df,
  split,
  caption = "Summary Statistics",
  total = TRUE,
  test = TRUE,
  na.rm = FALSE,
  general_note = NULL,
  p_note = "apa",
  max_width_in = 6,
  digits = 2
)
}
\arguments{
\item{df}{REQUIRED: Data frame}

\item{split}{REQUIRED: Bare varaiable name}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{total}{Optional: Logical. Defaults to add a total column}

\item{test}{Optional: Logical. Defaults to test groups differences}

\item{na.rm}{Optional: Logical. Defaults to NOT list-wise delete cases}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{max_width_in}{Optional: Number.  Inches wide table can be}

\item{digits}{Optional: Number. Digits after the decimal place}
}
\description{
Convert a \code{furniture::table1()} to a \code{flextable::flextable()} with APA formatting
}
\examples{
library(tidyverse)

# Simple
mtcars

}
