% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_apa.R
\name{theme_apa}
\alias{theme_apa}
\title{APA Formatted Table of a data.frame}
\usage{
theme_apa(tab, valign = "center", digits = 2, fontname = "serif", space = 0.5)
}
\arguments{
\item{tab}{A pre-created \code{flextable}}

\item{valign}{Optional: Text. vertical alignment ("center")}

\item{digits}{Optional: Number. Digits after the decimal place}

\item{fontname}{Optional: Text.  Font used in table}

\item{space}{Optional: Number. Line spacing in the body of the table}
}
\value{
table
}
\description{
APA Formatted Table of a data.frame
}
\examples{
library(tidyverse)
library(flextable)

mtcars \%>\%
  flextable::flextable() \%>\%
  theme_apa()
}
