% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaTab.R, R/gt_apa.R
\name{apaTab}
\alias{apaTab}
\title{APA Formatted Table of a data.frame}
\usage{
apaTab(
  x,
  caption,
  general_note = NULL,
  p_note = "* p < .05, ** p < .01, *** p < .001"
)

apaTab(
  x,
  caption,
  general_note = NULL,
  p_note = "* p < .05, ** p < .01, *** p < .001"
)
}
\arguments{
\item{x}{a data frame}

\item{caption}{text for the caption}

\item{general_note}{text for a general note}

\item{p_note}{text for a p-value significance indicators}

\item{note}{text for a general note}

\item{sig}{text for a p-value significance indicators}
}
\value{
table

table
}
\description{
APA Formatted Table of a data.frame

APA Formatted Table of a data.frame
}
\examples{
df <- head(cars)
apaTab(df, caption = "Testing")
df <- head(cars)
apaTab(df, caption = "Testing")
}
