% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_glm.R
\name{gt_glm}
\alias{gt_glm}
\title{APA: gtsummary for a single linear model}
\usage{
gt_glm(x, narrow = FALSE, fit = NA, d = 2, show_single_row = NULL)
}
\arguments{
\item{x}{REQUIRED: bare name. a single 'glm' object'}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: vector. quoted names of fit statistics to include, can be: "r.squared", "adj.r.squared", "sigma", "statistic","p.value", "df", "logLik", "AIC", "BIC", "deviance", "df.residual", and "nobs"}

\item{d}{Optional: number. digits after the decimal, default = 2}

\item{show_single_row}{(tidy-select) By default categorical variables are printed on multiple rows. If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}
}
\value{
a gtsummary object
}
\description{
APA: gtsummary for a single linear model
}
\examples{

library(tidyverse)

mtcars <- mtcars \%>\% dplyr::mutate(cyl = factor(cyl))

fit_glm1 <- glm(vs ~ wt, data = mtcars, family = "binomial")
fit_glm2 <- glm(vs ~ wt + mpg , data = mtcars, family = "binomial")

apaSupp::gt_glm(fit_glm2)
apaSupp::gt_glm(fit_glm2, narrow = TRUE, d = 3)
apaSupp::gt_glm(fit_glm2, narrow = TRUE, fit = c("AIC", "BIC"))

}
