% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_lm.R
\name{gt_lm}
\alias{gt_lm}
\title{APA: gtsummary for a single linear model}
\usage{
gt_lm(
  x,
  narrow = FALSE,
  fit = c("r.squared", "adj.r.squared"),
  d = 2,
  show_single_row = NULL
)
}
\arguments{
\item{x}{REQUIRED: bare name. a single 'lm' object'}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{fit}{Optional: Character vector. Default  = c("r.squared", "adj.r.squared").  May include quoted names of fit statistics listed from broom::glimpse(x)}

\item{d}{Optional: number. digits after the decimal, default = 2}

\item{show_single_row}{(tidy-select) By default categorical variables are printed on multiple rows. If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}
}
\value{
a gtsummary object
}
\description{
APA: gtsummary for a single linear model
}
\examples{

library(tidyverse)

mtcars <- mtcars \%>\% dplyr::mutate(vs = factor(vs))

m <- lm(mpg ~ vs + disp, data = mtcars)

apaSupp::gt_lm(m)
apaSupp::gt_lm(m, show_single_row = "vs")
apaSupp::gt_lm(m, d = 3)
apaSupp::gt_lm(m, narrow = TRUE)

}
