% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_cor.R
\name{tab_cor}
\alias{tab_cor}
\title{Tabulate Pairwise Pearson's Correlation}
\usage{
tab_cor(
  df,
  caption = "Correlations",
  general_note = NULL,
  p_note = "* p < .05. ** p < .01. *** p < .001.",
  max_width_in = 6,
  digits = 2
)
}
\arguments{
\item{df}{REQUIRED: A dataframe with selected numeric variables}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{max_width_in}{= Optional: Number.  Inches wide take can be}

\item{digits}{Optional: Number. Digits after the decimal place}
}
\value{
table
}
\description{
Tabulate Pairwise Pearson's Correlation
}
\examples{
library(tidyverse)
library(flextable)

mtcars

}
