% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_desc.R
\name{tab_desc}
\alias{tab_desc}
\title{Basic Descriptive Summary: n, M (SD), Median, etc. with flextable}
\usage{
tab_desc(
  df,
  total = FALSE,
  caption = "Descriptive Summary",
  general_note = NULL,
  p_note = NULL,
  valign = "center",
  digits = 2,
  fontname = "serif",
  space = 0.5
)
}
\arguments{
\item{df}{REQUIRED: data frame}

\item{total}{Optional: Logical. Calculate total? default is no (FALSE)}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. Significance note for APA table}

\item{valign}{Optional: Text. vertical alignment ("center")}

\item{digits}{Optional: Number. Digits after the decimal place}

\item{fontname}{Optional: Text.  Font used in table}

\item{space}{Optional: Number. Line spacing in the body of the table}
}
\value{
a \code{flextable} table with caption
}
\description{
Basic Descriptive Summary: n, M (SD), Median, etc. with flextable
}
\examples{
library(tidyverse)

# Simple
tab_desc(cars)

# Complex
cars \%>\%
    dplyr::rename("Speed, mph" = speed,
                  "Stopping Distance, ft" = dist) \%>\%
    tab_desc(caption = "Speed and Stopping Distances of Cars in the 1920's",
             general_note = "McNeil, D. R. (1977) Interactive Data Analysis. Wiley.")


}
