% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_desc.R
\name{tab_desc}
\alias{tab_desc}
\title{Basic Descriptive Summary: n, M (SD), \link{Min, Max}, Median with flextable}
\usage{
tab_desc(
  df,
  caption = "Descriptive Summary",
  general_note = NULL,
  p_note = NULL,
  fontname = "Times New Roman",
  align = "center",
  valign = "center",
  digits = 2,
  space = 1,
  ...
)
}
\arguments{
\item{df}{a data frame with relevant variables selected}

\item{caption}{the \code{flextable} caption for the table}

\item{general_note}{general note for footer of APA table}

\item{p_note}{significance note for APA table}

\item{fontname}{name of font for table}

\item{align}{alignment}

\item{valign}{vertial alignment}

\item{digits}{digits after the decimal place for numeric values}

\item{space}{line spacing in the body of the table, default is single}

\item{...}{extra arguments}
}
\value{
a \code{flextable} table with caption
}
\description{
Basic Descriptive Summary: n, M (SD), \link{Min, Max}, Median with flextable
}
\examples{
library(tidyverse)

tab_desc(cars)

cars \%>\%
dplyr::rename("Speed, mph" = speed,
              "Stopping Distance, ft" = dist) \%>\%
tab_desc(caption = "Speed and Stopping Distances of Cars in the 1920's",
         general_note = "McNeil, D. R. (1977) Interactive Data Analysis. Wiley.")


}
