% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_desc.R
\name{tab_desc}
\alias{tab_desc}
\title{Basic Descriptive Summary: n, M (SD), Median, etc. with flextable}
\usage{
tab_desc(
  df,
  caption = "Summary of Quantiatative Variables",
  general_note = NA,
  no_notes = FALSE,
  d = 2,
  max_width_in = 6
)
}
\arguments{
\item{df}{REQUIRED: Data frame}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{genderal_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{max_width_in}{= Optional: Number.  Inches wide the table can be}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}
}
\value{
a \code{flextable} table with caption
}
\description{
Basic Descriptive Summary: n, M (SD), Median, etc. with flextable
}
\examples{
library(tidyverse)

# Simple
tab_desc(cars)

# Complex
cars \%>\%
    dplyr::rename("Speed, mph" = speed,
                  "Stopping Distance, ft" = dist) \%>\%
    tab_desc(caption = "Speed and Stopping Distances of Cars in the 1920's",
             general_note = "McNeil, D. R. (1977) Interactive Data Analysis. Wiley.")


}
