% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_desc.R
\name{tab_desc}
\alias{tab_desc}
\title{Basic Descriptive Summary: n, M (SD), Median, etc. with flextable}
\usage{
tab_desc(
  df,
  caption = "Summary of Quantiatative Variables",
  docx = NA,
  tab_width = 0.9,
  general_note = NA,
  d = 2
)
}
\arguments{
\item{df}{REQUIRED: Data frame}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{tab_width}{Optional: numberic value (default is .9) \% of available width}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{d}{Optional: Number. Digits after the decimal place}
}
\value{
a \code{flextable} table with caption
}
\description{
Basic Descriptive Summary: n, M (SD), Median, etc. with flextable
}
\examples{
library(tidyverse)

# Simple
tab_desc(cars)

# Complex
cars \%>\%
    dplyr::rename("Speed, mph" = speed,
                  "Stopping Distance, ft" = dist) \%>\%
    tab_desc(caption = "Speed and Stopping Distances of Cars in the 1920's",
             general_note = "McNeil, D. R. (1977) Interactive Data Analysis. Wiley.")


}
