% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_freq.R
\name{tab_freq}
\alias{tab_freq}
\title{Frequency Table, total or split by a factor}
\usage{
tab_freq(
  df,
  split = NULL,
  caption = "Summary of Categorical Variables",
  docx = NA,
  tab_width = 0.9,
  general_note = NA,
  d = 2
)
}
\arguments{
\item{df}{REQUIRED: Data frame}

\item{split}{Optional: Quoted variable name}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{tab_width}{Optional: numberic value (default is .9) \% of available width}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{d}{Optional: Number. Digits after the decimal place}
}
\value{
a \code{flextable} table with caption
}
\description{
Frequency Table, total or split by a factor
}
\examples{
library(tidyverse)
library(forcats)

data(mtcars)

mtcars <- mtcars \%>\%
  dplyr::mutate(vs = factor(vs,
                            levels = 0:1,
                            labels = c("V Shaped", "Straight"))) \%>\%
  dplyr::mutate(am = factor(am,
                            levels = 0:1,
                            labels = c("Automatic", "Manual"))) \%>\%
  dplyr::mutate(cyl  = factor(cyl)) \%>\%
  dplyr::mutate(gear = factor(gear))


tab_freq(mtcars)

mtcars \%>\%
  dplyr::select(cyl,
                "Engine Type" = vs,
                "Transmission" = am,
                "Forward Gears" = gear) \%>\%
  tab_freq(split = "cyl",
           caption = "Summary of Automobile Engines by Number of Cylinders",
           general_note = "Data from the 1974 Motor Trend US magazine.")

}
