% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_glm.R
\name{tab_glm}
\alias{tab_glm}
\title{APA: flextable for a GLM models (specifically logisitc right now)}
\usage{
tab_glm(
  x,
  var_labels = NULL,
  caption = "Parameter Estimates for Generalized Linear Regression",
  general_note = NA,
  p_note = "apa123",
  no_notes = FALSE,
  d = 2,
  fit = NULL,
  pr2 = "tjur",
  vif = TRUE,
  lrt = TRUE,
  show_single_row = NULL,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***")
)
}
\arguments{
\item{x}{REQUIRED: a glm models, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa123"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{fit}{Optional: Text. fit statistics: (default = NA) "nobs", null.deviance", "df.null", "deviance", "df.residual", "logLik", "AIC", "BIC"}

\item{pr2}{Optional: character.  (default = "both") Include pseudo R-squared: "tjur", "mcfadden", "both", or "none" for logistic or "nagelkerke" for poisson}

\item{vif}{Optional: Logical. (default = TRUE) Include variance inflation factors}

\item{lrt}{Optional: Logical. (default = TRUE) Include LRT for single-predictor deletion}

\item{show_single_row}{a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}
}
\value{
a flextable object
}
\description{
APA: flextable for a GLM models (specifically logisitc right now)
}
\examples{

library(tidyverse)

mtcars <- mtcars \%>\% dplyr::mutate(am = factor(am))

fit_glm1 <- glm(vs ~ am, data = mtcars, family = "binomial")
fit_glm2 <- glm(vs ~ wt + mpg + am, data = mtcars, family = "binomial")

apaSupp::tab_glm(fit_glm1)
apaSupp::tab_glm(fit_glm2)

apaSupp::tab_glm(fit_glm2, vif = FALSE)
apaSupp::tab_glm(fit_glm2, lrt = FALSE)
apaSupp::tab_glm(fit_glm2, vif = FALSE, lrt = FALSE)

apaSupp::tab_glm(fit_glm2, pr2 = "both")
apaSupp::tab_glm(fit_glm2, pr2 = "tjur")
apaSupp::tab_glm(fit_glm2, pr2 = "mcfadden")
apaSupp::tab_glm(fit_glm2, pr2 = "none")


apaSupp::tab_glm(fit_glm2, fit = c("AIC", "BIC"))
apaSupp::tab_glm(fit_glm2, pr2 = "both", fit = c("AIC", "BIC"))


}
