% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_glm.R
\name{tab_glm}
\alias{tab_glm}
\title{APA: flextable for a GLM models (specifically logisitc right now)}
\usage{
tab_glm(
  x,
  var_labels = NULL,
  caption = "Generalized Regression Model",
  p_note = "apa",
  general_note = NA,
  fit = NULL,
  pr2 = "tjur",
  d = 2,
  vif = TRUE,
  lrt = TRUE,
  ...
)
}
\arguments{
\item{x}{REQUIRED: a glm models, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If p_note = "apa" then the standard "* p < .05. ** p < .01. *** p < .001." will be used}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{fit}{Optional: Text. fit statistics: (default = NA) "nobs", null.deviance", "df.null", "deviance", "df.residual", "logLik", "AIC", "BIC"}

\item{pr2}{Optional: character.  (default = "both") Include pseudo R-squared: "tjur", "mcfadden", "both", or "none"}

\item{d}{Optional: Number. Digits after the decimal place}

\item{vif}{Optional: Logical. (default = TRUE) Include variance inflation factors}

\item{lrt}{Optional: Logical. (default = TRUE) Include LRT for single-predictor deletion}

\item{type}{Optional: default = "logistic", more to come soon...}
}
\value{
a flextable object
}
\description{
APA: flextable for a GLM models (specifically logisitc right now)
}
\examples{

library(tidyverse)

mtcars <- mtcars \%>\% dplyr::mutate(am = factor(am))

fit_glm1 <- glm(vs ~ am, data = mtcars, family = "binomial")
fit_glm2 <- glm(vs ~ wt + mpg + am, data = mtcars, family = "binomial")

apaSupp::tab_glm(fit_glm1)
apaSupp::tab_glm(fit_glm2)

apaSupp::tab_glm(fit_glm2, vif = FALSE)
apaSupp::tab_glm(fit_glm2, lrt = FALSE)
apaSupp::tab_glm(fit_glm2, vif = FALSE, lrt = FALSE)

apaSupp::tab_glm(fit_glm2, pr2 = "both")
apaSupp::tab_glm(fit_glm2, pr2 = "tjur")
apaSupp::tab_glm(fit_glm2, pr2 = "mcfadden")
apaSupp::tab_glm(fit_glm2, pr2 = "none")


apaSupp::tab_glm(fit_glm2, fit = c("AIC", "BIC"))
apaSupp::tab_glm(fit_glm2, pr2 = "both", fit = c("AIC", "BIC"))


}
