% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lm.R
\name{tab_lm}
\alias{tab_lm}
\title{APA: flextable for ONE linear models}
\usage{
tab_lm(
  x,
  var_labels = NULL,
  caption = "Parameter Estimates for Linear Regression",
  docx = NA,
  general_note = NA,
  p_note = "apa123",
  no_notes = FALSE,
  d = 2,
  fit = c("r.squared", "adj.r.squared"),
  std = TRUE,
  vif = FALSE,
  eta2 = TRUE,
  ci = FALSE,
  show_single_row = NULL,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***")
)
}
\arguments{
\item{x}{REQUIRED: 1 lm model, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{docx}{Optional: filename. must end with ".docx"}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa123"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{general_note} and \code{p_note}}

\item{d}{Optional: Number. Digits after the decimal place}

\item{fit}{Optional: vector. quoted names of fit statistics to include, can be: "r.squared", "adj.r.squared", "sigma", "statistic","p.value", "df", "logLik", "AIC", "BIC", "deviance", "df.residual", and "nobs"}

\item{std}{Optional: Logical. (default = TRUE) Include standardized coefficients?}

\item{vif}{Optional: Logical. (default = FALSE) Include variance inflation factors?}

\item{eta2}{Optional: logical. (default = TRUE) Include eta-squared (semi-partial correlations) and partial eta-squared (partial correlations)}

\item{ci}{Optional: logical. (default = FALSE) Include a confidence interval for the estimated beta}

\item{show_single_row}{Optional: If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}
}
\value{
a flextable object
}
\description{
APA: flextable for ONE linear models
}
\examples{

m <- lm(mpg ~ disp + wt + hp, mtcars)

apaSupp::tab_lm(m)



}
