% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lm.R
\name{tab_lm}
\alias{tab_lm}
\title{APA: flextable for ONElinear models}
\usage{
tab_lm(
  model,
  var_labels = NULL,
  caption = "Regression Model",
  p_note = "apa",
  general_note = NULL,
  fit = c("r.squared", "adj.r.squared"),
  d = 2,
  std = TRUE,
  vif = FALSE,
  eta2 = TRUE
)
}
\arguments{
\item{model}{REQUIRED: 1 lm model, bare name}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If p_note = "apa" then the standard "* p < .05. ** p < .01. *** p < .001." will be used}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{fit}{Optional: vector. quoted names of fit statistics to include, can be: "r.squared", "adj.r.squared", "sigma", "statistic","p.value", "df", "logLik", "AIC", "BIC", "deviance", "df.residual", and "nobs"}

\item{d}{Optional: Number. Digits after the decimal place}

\item{std}{Optional: Logical. (default = TRUE) Include standardized coefficients?}

\item{vif}{Optional: Logical. (default = FALSE) Include variance inflation factors?}

\item{eta2}{Optional: logical. (default = TRUE) Include eta-squared (semi-partial correlations) and partial eta-squared (partial correlations)}
}
\value{
a flextable object
}
\description{
APA: flextable for ONElinear models
}
\examples{

m <- lm(dist ~ speed, cars)
tab_lm(m)

}
