% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lms.R
\name{tab_lms}
\alias{tab_lms}
\title{APA: flextable for 2-3 linear models}
\usage{
tab_lms(
  x,
  var_labels = NULL,
  caption = "Compare Regression Models",
  narrow = FALSE,
  p_note = "apa",
  general_note = NULL,
  fit = c("r.squared", "adj.r.squared"),
  d = 2,
  show_single_row = NULL
)
}
\arguments{
\item{x}{REQUIRED: List. at least 2 lm models, bare names}

\item{var_labels}{Optional: Vector. Text replacements for model terms, "old" = "new"}

\item{caption}{Optional: Text. Caption for the table}

\item{narrow}{Optional. Logical. Default = FALSE, but TRUE will exclude p-vlaues from the table to make it narrower}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If p_note = "apa" then the standard "* p < .05. ** p < .01. *** p < .001." will be used}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{fit}{Optional: vector. quoted names of fit statistics to include, can be: "r.squared", "adj.r.squared", "sigma", "statistic","p.value", "df", "logLik", "AIC", "BIC", "deviance", "df.residual", and "nobs"}

\item{d}{Optional: Number. Digits after the decimal place}

\item{show_single_row}{(tidy-select) By default categorical variables are printed on multiple rows. If a variable is dichotomous (e.g. Yes/No) and you wish to print the regression coefficient on a single row, include the variable name(s) here.}
}
\value{
a flextable object
}
\description{
APA: flextable for 2-3 linear models
}
\examples{

m1 <- lm(dist ~ 1, cars)
m2 <- lm(dist ~ speed, cars)

tab_lms(list(m1, m2))

}
