% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_xalpha.R
\name{tab_xalpha}
\alias{tab_xalpha}
\title{Tabulate Internal Consistency}
\usage{
tab_xalpha(
  df,
  measure = NULL,
  domain = NULL,
  item,
  value,
  id,
  caption = "Internal Consistency",
  general_note = NA,
  d = 2,
  max_width_in = 6
)
}
\arguments{
\item{df}{REQUIRED: data frame}

\item{measure}{Optional: Bare variable name. Measurement tool name}

\item{domain}{Optional: Bare variable name. Domain or sub-score name}

\item{item}{REQUIRED: Bare variable name. Item distinction}

\item{value}{REQUIRED: Bare variable name. Values or rating of each item}

\item{id}{REQUIRED: Bare variable name. Identify cases/subjects}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{d}{Optional: Number. Digits after the decimal place}

\item{max_width_in}{= Optional: Number.  Inches wide the table can be}
}
\value{
a \code{flextable} table with caption
}
\description{
Tabulate Internal Consistency
}
\examples{
library(tidyverse)
library(lavaan)

data(HolzingerSwineford1939, package = "lavaan")

HolzingerSwineford1939 \%>\%
  tidyr::pivot_longer(cols = starts_with("x"),
             names_to = "it",
             values_to = "val") \%>\%
  dplyr::mutate(Domain = case_when(it \%in\% c("x1", "x2", "x3") ~ "Visual",
                                   it \%in\% c("x4", "x5", "x6") ~ "Verbal",
                                  it \%in\% c("x7", "x8", "x9") ~ "Speed")) \%>\%
  dplyr::mutate(Measure = case_when(Domain \%in\% c("Visual", "Verbal") ~
                                   "Scale V Made Up",
                                    Domain == "Speed"  ~
                                   "Scale S With A Very Long Name for Demo")) \%>\%
  tab_xalpha(measure = Measure,
             domain = Domain,
             item = it,
             value = val,
             id = id,
             caption = "Internal Consistency for the Holzinger and Swineford",
             general_note = "Scales are for example only.")

}
