% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_apa.R
\name{theme_apa}
\alias{theme_apa}
\title{APA Formatted Table of a data.frame}
\usage{
theme_apa(
  x,
  caption = "Table Caption",
  general_note = NA,
  p_note = NA,
  no_notes = FALSE,
  breaks = c(0.05, 0.01, 0.001),
  symbols = c("*", "**", "***"),
  d = 2,
  max_width_in = 6,
  main_note = NA
)
}
\arguments{
\item{x}{REQUIRED: A pre-created \code{flextable} object}

\item{caption}{REQUIRED: Text. Caption for the table}

\item{general_note}{Optional: Text. General note for footer of APA table}

\item{p_note}{Optional: Text. (default = NULL) Significance note for APA table, If \code{p_note = "apa123"} then the standard \code{"* p < .05. ** p < .01. *** p < .001."} will be used}

\item{no_notes}{REQUIRED: Logical.  Defaults to \code{FALSE}, if \code{TRUE} will ignore \code{genderal_note} and \code{p_note}}

\item{breaks}{Optional: numeric vector of p-value cut-points}

\item{symbols}{Optional: character vector for symbols denoting p-value cut-points}

\item{d}{Optional: Number. Digits after the decimal place}

\item{max_width_in}{Optional: Number.  Inches wide take can be}

\item{main_note}{Optional: alternative to \code{general_note} already in paragraph form}
}
\value{
table
}
\description{
APA Formatted Table of a data.frame
}
\examples{
library(tidyverse)
library(flextable)

mtcars \%>\%
  dplyr::select(mpg, disp, hp, drat, wt) \%>\%
  flextable::flextable() \%>\%
  theme_apa(caption = "Summary of Some Variables")
}
