% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_to_tibble.R
\name{vif_to_tibble}
\alias{vif_to_tibble}
\title{APA: calculate VIF}
\usage{
vif_to_tibble(model, d = 2)
}
\arguments{
\item{d}{Optional: Number. Digits after the decimal place}

\item{x}{REQUIRED: a models (lm or glm), bare name}
}
\value{
a data.frame for use with apaSupp's model tables
}
\description{
APA: calculate VIF
}
\examples{

mtcars <- mtcars \%>\% dplyr::mutate(cyl = factor(cyl))

fit_lm1 <- lm(mpg ~ wt + hp + cyl, data = mtcars)

vif_to_tibble(fit_lm1)

}
